+++ src/main/kotlin/blackjack/domain/blackjackgame/GameStatus.kt
@@ -1,8 +1,8 @@
 package blackjack.domain.blackjackgame
 
 import blackjack.domain.card.PlayerDeck
-import blackjack.judge.Judgement
-import blackjack.state.Ready
-import blackjack.state.State
+import blackjack.domain.judge.Judgement
+import blackjack.domain.state.Ready
+import blackjack.domain.state.State
 
 class GameStatus(val state: State = Ready(PlayerDeck()), val judgements: MutableList<Judgement> = mutableListOf())
