+++ src/main/kotlin/blackjack/domain/judge/Judgement.kt
@@ -1,0 +1,3 @@
+package blackjack.domain.judge
+
+interface Judgement
