+++ src/main/kotlin/blackjack/domain/judge/Win.kt
@@ -1,4 +1,4 @@
-package blackjack.judge
+package blackjack.domain.judge
 
 class Win : Judgement {
