--- src/main/kotlin/blackjack/judge/Judgement.kt
@@ -1,3 +1,0 @@
-package blackjack.judge
-
-interface Judgement
