+++ src/test/kotlin/blackjack/domain/player/DealerTest.kt
@@ -3,7 +3,7 @@ package blackjack.domain.player
 import blackjack.domain.card.Card
 import blackjack.domain.card.CardShape
 import blackjack.domain.card.CardSymbol
-import blackjack.state.Ready
+import blackjack.domain.state.Ready
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
 import org.junit.jupiter.api.Test
