+++ src/test/kotlin/blackjack/state/HitTest.kt
@@ -4,6 +4,7 @@ import blackjack.domain.card.Card
 import blackjack.domain.card.CardShape
 import blackjack.domain.card.CardSymbol
 import blackjack.domain.card.PlayerDeck
+import blackjack.domain.state.Hit
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.Test
